Rename browser_detection file into env_detector
