+++ src/compat/__tests__/can_rely_on_video_visibility_and_size.test.ts
@@ -1,6 +1,6 @@
 import { describe, afterEach, it, expect, vi } from "vitest";
-import type IEnvDetector from "../browser_detection";
 import type ICanRelyOnVideoVisibilityAndSize from "../can_rely_on_video_visibility_and_size";
+import type IEnvDetector from "../env_detector";
 
 describe("Compat - canRelyOnVideoVisibilityAndSize", () => {
   afterEach(() => {
@@ -8,8 +8,8 @@ describe("Compat - canRelyOnVideoVisibilityAndSize", () => {
   });
 
   it("should return true on any browser but Firefox", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
@@ -25,8 +25,8 @@ describe("Compat - canRelyOnVideoVisibilityAndSize", () => {
   });
 
   it("should return true on Firefox but the version is unknown", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
@@ -45,8 +45,8 @@ describe("Compat - canRelyOnVideoVisibilityAndSize", () => {
   });
 
   it("should return true on Firefox < 67>", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
@@ -65,8 +65,8 @@ describe("Compat - canRelyOnVideoVisibilityAndSize", () => {
   });
 
   it("should return false on Firefox >= 67", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
