+++ src/compat/__tests__/should_wait_for_data_before_loaded.test.ts
@@ -1,5 +1,5 @@
 import { describe, beforeEach, it, expect, vi } from "vitest";
-import type IEnvDetector from "../browser_detection";
+import type IEnvDetector from "../env_detector";
 import type IShouldWaitForDataBeforeLoaded from "../should_wait_for_data_before_loaded";
 
 describe("compat - shouldWaitForDataBeforeLoaded", () => {
@@ -8,8 +8,8 @@ describe("compat - shouldWaitForDataBeforeLoaded", () => {
   });
 
   it("should return true if we are not on Safari browser nor in directfile mode", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
@@ -25,8 +25,8 @@ describe("compat - shouldWaitForDataBeforeLoaded", () => {
   });
 
   it("should return true if we are not on Safari browser but in directfile mode", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
@@ -42,8 +42,8 @@ describe("compat - shouldWaitForDataBeforeLoaded", () => {
   });
 
   it("should return true if we are on the Safari mobile browser but not in directfile mode", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
@@ -59,8 +59,8 @@ describe("compat - shouldWaitForDataBeforeLoaded", () => {
   });
 
   it("should return true if we are on the Safari desktop browser but not in directfile mode", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
@@ -76,8 +76,8 @@ describe("compat - shouldWaitForDataBeforeLoaded", () => {
   });
 
   it("should return false if we are on the Safari mobile browser and in directfile mode", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
@@ -93,8 +93,8 @@ describe("compat - shouldWaitForDataBeforeLoaded", () => {
   });
 
   it("should return false if we are on the Safari desktop browser and in directfile mode", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
