+++ src/compat/__tests__/should_wait_for_have_enough_data.test.ts
@@ -1,5 +1,5 @@
 import { describe, beforeEach, it, expect, vi } from "vitest";
-import type IEnvDetector from "../browser_detection";
+import type IEnvDetector from "../env_detector";
 import type IShouldWaitForHaveEnoughData from "../should_wait_for_have_enough_data";
 
 describe("compat - shouldWaitForHaveEnoughData", () => {
@@ -8,8 +8,8 @@ describe("compat - shouldWaitForHaveEnoughData", () => {
   });
 
   it("should return false if we are not on the Playstation 5", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
@@ -26,8 +26,8 @@ describe("compat - shouldWaitForHaveEnoughData", () => {
   });
 
   it("should return true if we are on the Playstation 5", async () => {
-    vi.doMock("../browser_detection", async () => {
-      const EnvDetector = (await vi.importActual("../browser_detection"))
+    vi.doMock("../env_detector", async () => {
+      const EnvDetector = (await vi.importActual("../env_detector"))
         .default as typeof IEnvDetector;
       return {
         default: {
