+++ src/compat/env_detector.ts
