+++ src/compat/has_issues_with_high_media_source_duration.ts
@@ -1,4 +1,4 @@
-import EnvDetector from "./browser_detection";
+import EnvDetector from "./env_detector";
 
 /**
  * Some platforms have issues when the `MediaSource`'s `duration` property
