+++ src/compat/may_media_element_fail_on_undecipherable_data.ts
@@ -1,4 +1,4 @@
-import EnvDetector from "./browser_detection";
+import EnvDetector from "./env_detector";
 
 /**
  * We noticed that the PlayStation 5 may have the HTMLMediaElement on which the
