+++ src/compat/remove_cue.ts
@@ -16,7 +16,7 @@
 
 import log from "../log";
 import type { ICompatTextTrack, ICompatVTTCue } from "./browser_compatibility_types";
-import EnvDetector from "./browser_detection";
+import EnvDetector from "./env_detector";
 
 /**
  * Return true if given cue is active.
