+++ src/compat/should_validate_metadata.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import EnvDetector from "./browser_detection";
+import EnvDetector from "./env_detector";
 
 /**
  * Returns true if the metadata received after a "loadedmetadata" event has
