+++ src/compat/should_wait_for_data_before_loaded.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import EnvDetector from "./browser_detection";
+import EnvDetector from "./env_detector";
 
 /**
  * On some browsers, the ready state might never go above `1` when autoplay is
